apps/data-seeding/requirements.txt
sarafu-faucet~=0.0.7
cic-eth[tools]~=0.12.8
cic-types~=0.2.1a8
funga~=0.5.1
faker==4.17.1
chainsyncer~=0.1.0
chainlib-eth~=0.0.19
eth-address-index~=0.2.4
eth-contract-registry~=0.7.2
eth-accounts-index~=0.1.2
eth-erc20==0.1.6
erc20-faucet~=0.3.2
psycopg2==2.8.6
liveness~=0.0.1a7
confini~=0.5.3
leveldir~=0.2.0
